implemented delete_post, and now id in JWT is int64

+++ biz/handler/base/verification_service.go
@@ -152,7 +152,7 @@ func Me(ctx context.Context, c *app.RequestContext) {
 
 	c.JSON(consts.StatusOK, verification.MeResp{
 		IsSuccessful: true,
-		ID:           int32(id),
+		ID:           id,
 		Username:     username,
 		Email:        email,
 	})
